test: Combine instrinsics tests into a single test file
